+++ scripts/components/toggle/docs/readme.mdx
@@ -3,14 +3,14 @@ import * as ComponentData from './story.js';
 
 <Meta of={ComponentData} name='About' />
 
-# IconToggle
+# Toggle
 
 A toggle/checkbox control with an optional icon.
 
 ## Usage
 
 ```jsx
-<IconToggle
+<Toggle
 	icon={icons.width}
 	label={__('Wide button', 'eightshift-frontend-libs')}
 	checked={buttonFullWidth}
