+++ scripts/components/toggle/toggle.js
@@ -1,11 +1,11 @@
 import React from 'react';
 import { ToggleControl, CheckboxControl, Button } from '@wordpress/components';
-import { IconLabel, classnames, TileButton } from '../../../scripts';
+import { IconLabel, classnames, TileButton } from '../..';
 
 /**
- * Custom toggle control that can render as
+ * Custom toggle control.
  *
- * @param {object} props                                                                 - IconToggle options.
+ * @param {object} props                                                                 - `Toggle` options.
  * @param {React.Component} props.icon                                                   - Icon to display.
  * @param {string} props.label                                                           - Usually component name.
  * @param {'toggle'|'checkbox'|'button'|'iconButton'|'tileButton'} [props.type='toggle'] - Kind of toggle to render.
@@ -18,7 +18,7 @@ import { IconLabel, classnames, TileButton } from '../../../scripts';
  * @param {boolean?} [props.reducedBottomSpacing]                                        - If `true`, space below the control is reduced.
  * @param {string?} [props.additionalClasses]                                            - If provided, classes are passed to the underlying component.
  */
-export const IconToggle = ({
+export const Toggle = ({
 	icon,
 	label,
 
@@ -85,3 +85,10 @@ export const IconToggle = ({
 		/>
 	);
 };
+
+/**
+ * Custom toggle control.
+ *
+ * @deprecated since 9.5.0 - renamed to `Toggle`
+ */
+export const IconToggle = (props) => <Toggle {...props} />;
